// src/components/state/types/actionEnums.ts
export enum ActionEnums {
  AUTH_IS_LOADING = "AUTH_IS_LOADING",
  ADMIN_IS_LOADING = "ADMIN_IS_LOADING",
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAIL = "SIGN_IN_FAIL",
  SIGNOUT_SUCCESS = "SIGNOUT_SUCCESS",
  SIGNOUT_FAIL = "SIGNOUT_FAIL",
  SIGN_UP_SUCCESS = "SIGN_UP_SUCCESS",
  SIGN_UP_FAIL = "SIGN_UP_FAIL",
  GET_USER_SUCCESS = "GET_USER_SUCCESS",
  GET_USER_FAIL = "GET_USER_FAIL",
  GET_USER_DETAILS_SUCCESS = "GET_USER_DETAILS_SUCCESS",
  GET_USER_DETAILS_FAIL = "GET_USER_DETAILS_FAIL",
  ADMIN_GET_ALL_USERS_SUCCESS = "ADMIN_GET_ALL_USERS_SUCCESS",
  ADMIN_GET_ALL_USERS_FAIL = "ADMIN_GET_ALL_USERS_FAIL",
  ADMIN_UPDATE_USER_SUCCESS = "ADMIN_UPDATE_USER_SUCCESS",
  ADMIN_UPDATE_USER_FAIL = "ADMIN_UPDATE_USER_FAIL",
  ADMIN_DELETE_USER_SUCCESS = "ADMIN_DELETE_USER_SUCCESS",
  ADMIN_DELETE_USER_FAIL = "ADMIN_DELETE_USER_FAIL",
}